Task_1_Task_2_API_Automation/requirements.txt
requests==2.31.0
pytest==7.4.3
jsonschema==4.20.0
python-dotenv==1.0.0
